Make extract reusable by fetching secrets from env (#57)
By removing the credentials for feedbin-extract from the image
it can be reused regardless of set credentials. This also opens
the opportunity to distribute the image openly on Dockerhub.
It also seems like a better habit to not have them stored
in an image directly.

This change introduces an entrypoint for feedbin-extract
that will, before starting node, fetch the credentials
from env and store them in the container filesystem.